First commit
So far this program is ran with a connection string to connect to mongodb and from there it has a function that takes the client, first db, second db, first collection, second collection, and sample_size and compares the sample_size of documents from the first collection and sees if it exists exactly how it is in the second collection. Returning true with a print that says yes they are identical or else false.