Improved Functionality + Added Test Cases Checks
Added DeepDiff for telling when theres a difference and making it useful. Made the code rely less on storing docs and values in different lists when it was not needed and practical for large datasets. Additionally, added a progress bar so user is not left confused on how long it will run. Also made it change from passing in a sample_size to passing a percent that will round and convert to a sample size on its own. Lastly, implemented a bunch of test cases to be checked, if collection is empty, or differing in size, etc.