Updated code for passing all edge cases.
Removed Redundant check to see if sample size is 0. Next, formatted strings to make the code easier to read. Utilized dumps to compare strings to see if they are equal. Added js scripts that you can run on mongo to populate your databases. Did a find on the ID field not on the document as a whole, and did a find_one. Removed second internal for loop, it was needed in the dataCompare. Lastly, made the output more useful and readable to user.